% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawElement.R
\docType{methods}
\name{drawElement,input.button.download-method}
\alias{drawElement,input.button.download-method}
\title{Title}
\usage{
\S4method{drawElement}{input.button.download}(x)
}
\arguments{
\item{x}{input.button.download.}
}
\value{
返回值
}
\description{
Title
}
\examples{
drawElement(input.button.download());
}
