% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawElement.R
\docType{methods}
\name{drawElement,input.checkboxGroup-method}
\alias{drawElement,input.checkboxGroup-method}
\title{定义drawElement方法在input.checkboxGroup上的实现}
\usage{
\S4method{drawElement}{input.checkboxGroup}(x)
}
\arguments{
\item{x}{input.checkboxGroup. 实例}
}
\value{
返回值
}
\description{
定义drawElement方法在input.checkboxGroup上的实现
}
\examples{
drawElement();
}
