% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawElement.R
\docType{methods}
\name{drawElement,input.date-method}
\alias{drawElement,input.date-method}
\title{定义drawElement方法在input.date上的实现}
\usage{
\S4method{drawElement}{input.date}(x)
}
\arguments{
\item{x}{input.date. 实例化对象}
}
\value{
返回值
}
\description{
定义drawElement方法在input.date上的实现
}
\examples{
drawElement();
}
