% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawElement.R
\docType{methods}
\name{drawElement,input.file-method}
\alias{drawElement,input.file-method}
\title{定义drawElement方法在input.file上的实现}
\usage{
\S4method{drawElement}{input.file}(x)
}
\arguments{
\item{x}{input.file.  input.file类}
}
\value{
返回值
}
\description{
定义drawElement方法在input.file上的实现
}
\examples{
 drawElement(input.file());
}
