% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawElement.R
\docType{methods}
\name{drawElement,input.numeric-method}
\alias{drawElement,input.numeric-method}
\title{定义drawElement方法在input.numeric上的实现}
\usage{
\S4method{drawElement}{input.numeric}(x)
}
\arguments{
\item{x}{input.numeric. 定义对象}
}
\value{
返回值
}
\description{
定义drawElement方法在input.numeric上的实现
}
\examples{
drawElement();
}
