% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawElement.R
\docType{methods}
\name{drawElement,input.slider-method}
\alias{drawElement,input.slider-method}
\title{定义drawElement方法在input.slider类的实现}
\usage{
\S4method{drawElement}{input.slider}(x)
}
\arguments{
\item{x}{input.slider. 输入对象实例}
}
\value{
返回值
}
\description{
定义drawElement方法在input.slider类的实现
}
\examples{
drawElement();
}
