% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawElement.R
\docType{methods}
\name{drawElement,input.text-method}
\alias{drawElement,input.text-method}
\title{定义drawElement方法在input.text上的实现}
\usage{
\S4method{drawElement}{input.text}(x)
}
\arguments{
\item{x}{input.text. input.text类}
}
\value{
返回值
}
\description{
定义drawElement方法在input.text上的实现
}
\examples{
drawElement();
}
