% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawElement.R
\docType{methods}
\name{drawElement,output.dataTable-method}
\alias{drawElement,output.dataTable-method}
\title{定义drawElement方法在output.dataTable类的实现}
\usage{
\S4method{drawElement}{output.dataTable}(x)
}
\arguments{
\item{x}{output.dataTable. 类output.dataTable实例对象}
}
\value{
返回值
}
\description{
定义drawElement方法在output.dataTable类的实现
}
\examples{
drawElement(output.dataTable());
}
