% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawElement.R
\docType{methods}
\name{drawElement,output.image-method}
\alias{drawElement,output.image-method}
\title{Title}
\usage{
\S4method{drawElement}{output.image}(x)
}
\arguments{
\item{x}{output.image.  output.image类的对象}
}
\value{
返回值
}
\description{
Title
}
\examples{
drawElement();
}
