% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawElement.R
\docType{methods}
\name{drawElement,output.print-method}
\alias{drawElement,output.print-method}
\title{定义drawElement方法在output.print}
\usage{
\S4method{drawElement}{output.print}(x)
}
\arguments{
\item{x}{output.print. output.print实例}
}
\value{
返回值
}
\description{
定义drawElement方法在output.print
}
\examples{
drawElement();
}
