% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawElement.R
\name{drawElement}
\alias{drawElement}
\title{定义drawElement方法的泛型函数Interface}
\usage{
drawElement(x)
}
\arguments{
\item{x}{对象的参数，后续可能考虑进行扩展}
}
\value{
返回值
}
\description{
定义drawElement方法的泛型函数Interface
}
\examples{
drawElemnt();
}
