% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.image.R
\name{format_image}
\alias{format_image}
\title{Title image formmater}
\usage{
format_image(fileName = "demo.jpg", height = "400px",
  width = "400px", alt = "image file holderplace")
}
\arguments{
\item{fileName}{filename}

\item{height}{height}

\item{width}{width}

\item{alt}{alter text}
}
\value{
return_value
}
\description{
Title image formmater
}
\examples{
format.image();
}
