% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArgList.R
\name{getArgList}
\alias{getArgList}
\title{获取函数的参数列表}
\usage{
getArgList(func_names)
}
\arguments{
\item{func_names}{函数名称,character向量,支持批量处理}
}
\value{
返回一个数据框,提供FunctionName与FunctionArgList两个字段
}
\description{
获取函数的参数列表
}
\examples{
 getArgList(c('actionButton','actionLink'));
}
