% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.button.download.R
\docType{class}
\name{input.button.download-class}
\alias{input.button.download-class}
\title{定义input.button.download输入控件}
\value{
返回值
}
\description{
定义input.button.download输入控件
}
\section{Slots}{

\describe{
\item{\code{class}}{character. css类名称}
}}

\examples{
input.button.download()
}
