% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.button.download.R
\name{input.button.download}
\alias{input.button.download}
\title{定义input.button.download的同名的入口函数}
\usage{
input.button.download(Id = "downloadButton1", label = "下载文件",
  css_class = "downloadButton")
}
\arguments{
\item{Id}{内码}

\item{css_class}{css类名称}

\item{Label}{标签名称}
}
\value{
返回值
}
\description{
定义input.button.download的同名的入口函数
}
\examples{
input.button.download();
}
