% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.checkboxGroup.R
\docType{class}
\name{input.checkboxGroup-class}
\alias{input.checkboxGroup-class}
\title{定义input.checkboxGroup基类}
\value{
返回值
}
\description{
定义input.checkboxGroup基类
}
\examples{
input.checkboxGroup();
}
