% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.checkboxGroup.R
\name{input.checkboxGroup}
\alias{input.checkboxGroup}
\title{定义input.checkboxGroup同名的入口函数}
\usage{
input.checkboxGroup(Id = "checkboxGroupInput1", label = "请选择:",
  choiceNames = list(icon("calendar"), icon("bed"), icon("cog"),
  icon("bug")), choiceValues = list("calendar", "bed", "cog", "bug"),
  selected = "bug", inline = FALSE, width = "100\%")
}
\arguments{
\item{Id}{input内码}

\item{label}{标签}

\item{choiceNames}{选项名称}

\item{choiceValues}{选择值}

\item{selected}{默认初始化选项}

\item{inline}{是否横向展示}

\item{width}{宽度}
}
\value{
返回值
}
\description{
定义input.checkboxGroup同名的入口函数
}
\examples{
input.checkboxGroup();
}
