% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.date.R
\docType{class}
\name{input.date-class}
\alias{input.date-class}
\title{定义input.date类,可以实例化}
\value{
返回值
}
\description{
定义input.date类,可以实例化
}
\section{Slots}{

\describe{
\item{\code{value}}{Date. 默认值}
}}

\examples{
input.date();
}
