% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.date.R
\name{input.date}
\alias{input.date}
\title{定义input.date的同名入口函数}
\usage{
input.date(Id = "dateInput1", label = "请选择一个日期:",
  value = Sys.Date(), min = as.Date("1984-01-01"),
  max = as.Date("2100-12-31"), format = "yyyy-mm-dd",
  startview = "month", weekstart = 1L, language = "zh-CN",
  width = "100\%", autoclose = TRUE)
}
\arguments{
\item{Id}{inputId}

\item{label}{标签名称}

\item{value}{默认当前日期}

\item{min}{最小值}

\item{max}{最大值}

\item{format}{日期格式}

\item{startview}{日历格式}

\item{weekstart}{一周开始某哪一天}

\item{language}{语言默认为中文}

\item{width}{宽度}

\item{autoclose}{自动关闭}
}
\value{
返回值
}
\description{
定义input.date的同名入口函数
}
\examples{
input.date();
}
