% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.dateRange.R
\docType{class}
\name{input.dateRange-class}
\alias{input.dateRange-class}
\title{定义input.dataRange类，可以实例化}
\value{
返回值
}
\description{
定义input.dataRange类，可以实例化
}
\section{Slots}{

\describe{
\item{\code{start}}{Date.开始日期}

\item{\code{separator}}{分隔符}

\item{\code{end}}{Date.结束日期}
}}

\examples{
input.dateRange()
}
