% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.dateRange.R
\name{input.dateRange}
\alias{input.dateRange}
\title{定义input.dateRange同名实例化入口函数}
\usage{
input.dateRange(Id = "dateRangeInput1",
  label = "请选择日期范围:", start = Sys.Date(),
  end = Sys.Date() + 14, min = as.Date("1984-01-01"),
  max = as.Date("2100-12-31"), format = "yyyy-mm-dd",
  startview = "month", weekstart = 1L, language = "zh-CN",
  separator = "到", width = "100\%", autoclose = TRUE)
}
\arguments{
\item{Id}{inputId}

\item{label}{标签名称}

\item{start}{开始日期默认当天}

\item{end}{结束日期默认2周后}

\item{min}{最小日期}

\item{max}{最大日期}

\item{format}{日期格式}

\item{startview}{月历格式}

\item{weekstart}{一周开始于}

\item{language}{语言}

\item{separator}{分隔符号}

\item{width}{宽度}

\item{autoclose}{自动关闭}
}
\value{
返回值
}
\description{
定义input.dateRange同名实例化入口函数
}
\examples{
input.dateRange();
}
