% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.file.R
\docType{class}
\name{input.file-class}
\alias{input.file-class}
\title{定义input.file类,可以进行实例化}
\value{
返回值
}
\description{
定义input.file类,可以进行实例化
}
\section{Slots}{

\describe{
\item{\code{multiple}}{logical. 是否可以多选}

\item{\code{accept}}{character.  可以接受的文件类型}

\item{\code{buttonLabel}}{character. 默认的按纽标签文字}

\item{\code{placeholder}}{character. 占位符}
}}

\examples{
input.file();
}
