% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.file.R
\name{input.file}
\alias{input.file}
\title{定义input.file的同步入口函数}
\usage{
input.file(Id = "FileInput1", label = "上传文件1",
  width = "auto", multiple = FALSE, accept = "",
  buttonLabel = "请选择文件...",
  placeholder = "没有选任何文件")
}
\arguments{
\item{Id}{内码}

\item{label}{标签名称}

\item{width}{宽度}

\item{multiple}{是否可以多选}

\item{accept}{文件类型}

\item{buttonLabel}{默认按纽名称}

\item{placeholder}{默认提供选项}
}
\value{
返回值
}
\description{
定义input.file的同步入口函数
}
\examples{
input.file();
}
