% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.numeric.R
\name{input.numeric}
\alias{input.numeric}
\title{定义input.numeric的同名函数}
\usage{
input.numeric(Id = "numericInput1", label = "请输入数值:",
  value = 30, min = 5, max = 50, step = 1, width = "100\%")
}
\arguments{
\item{Id}{内码}

\item{label}{名称}

\item{value}{默认值}

\item{min}{最小值}

\item{max}{最大值}

\item{step}{步骤}
}
\value{
返回值
}
\description{
定义input.numeric的同名函数
}
\examples{
input.numeric();
}
