% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.slider.R
\docType{class}
\name{input.slider-class}
\alias{input.slider-class}
\title{定义input.slider的基类,可实例化}
\value{
返回值
}
\description{
定义input.slider的基类,可实例化
}
\section{Slots}{

\describe{
\item{\code{round}}{integer. 负二表示2位小数，正一表达十，与EXCEL正好相反的表达}
}}

\examples{
input.slider();
}
