% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.slider.R
\name{input.slider}
\alias{input.slider}
\title{定义input.slider同名的入口函数}
\usage{
input.slider(Id = "sliderInput1", label = "请滚动执行数据:",
  min = 0, max = 100, value = 60, step = 1, round = 0L,
  width = "100\%")
}
\arguments{
\item{Id}{input内码}

\item{label}{标签名称}

\item{min}{最小值下限}

\item{max}{最大值上限}

\item{value}{默认值}

\item{round}{四舍五入}
}
\value{
返回值
}
\description{
定义input.slider同名的入口函数
}
\examples{
input.slider();
}
