% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.text.R
\name{input.text}
\alias{input.text}
\title{定义文本或密码输入框的实例化类}
\usage{
input.text(Id = "textInput1", label = "请输入文本", value = "",
  width = "100\%", placeholder = "", isPassword = FALSE)
}
\arguments{
\item{Id}{内码}

\item{label}{标签名称}

\item{value}{默认值}

\item{width}{宽度}

\item{placeholder}{点位符}

\item{isPassword}{是否为密码,默认为否}
}
\value{
返回值
}
\description{
定义文本或密码输入框的实例化类
}
\examples{
input.text();
}
