% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl.button.download.R
\name{mdl.button.download}
\alias{mdl.button.download}
\title{定义下载UI组件mdl.button.download}
\usage{
mdl.button.download(id, label = "下载文件")
}
\arguments{
\item{id}{命名空间id}

\item{label}{下载按纽的名称}
}
\value{
返回值
}
\description{
定义下载UI组件mdl.button.download
}
\examples{
mdl.button.download();
}
