% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl.dataTable.preview.R
\name{mdl.dataTable.preview}
\alias{mdl.dataTable.preview}
\title{定义数据预览的逻辑}
\usage{
mdl.dataTable.preview(id, label)
}
\arguments{
\item{id}{预算组件的ID}

\item{label}{输入文件的标签}
}
\value{
返回值
}
\description{
定义数据预览的逻辑
}
\examples{
mdl.dataTable.preview();
}
