% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl.dataTable.preview.R
\name{mdl.dataTable.preview2}
\alias{mdl.dataTable.preview2}
\title{定义数据预览的逻辑2}
\usage{
mdl.dataTable.preview2(id, label)
}
\arguments{
\item{id}{组件id}

\item{label}{文件标签}
}
\value{
返回值
}
\description{
定义数据预览的逻辑2
}
\examples{
mdl.dataTable.preview2();
}
