% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl.image.R
\name{mdl.image}
\alias{mdl.image}
\title{mdl.image module function}
\usage{
mdl.image(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
return value
}
\description{
mdl.image module function
}
\examples{
mdl.image();
}
