% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl.numeric.R
\name{mdl.numeric}
\alias{mdl.numeric}
\title{定义numeric相应的组件}
\usage{
mdl.numeric(id, label = "请输入数值:")
}
\arguments{
\item{id}{内码}

\item{label}{名称}
}
\value{
返回值
}
\description{
定义numeric相应的组件
}
\examples{
mdl.numeric();
}
