% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.dataTable.R
\name{output.dataTable}
\alias{output.dataTable}
\title{定义output.dataTable的同名入口函数}
\usage{
output.dataTable(Id = "dataTableOutput1")
}
\arguments{
\item{Id}{outputId}
}
\value{
返回值
}
\description{
定义output.dataTable的同名入口函数
}
\examples{
output.dataTable();
}
