% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.plot.R
\name{output.plot}
\alias{output.plot}
\title{定义output.image实例化入口函数}
\usage{
output.plot(Id = "plotOutput1", width = "100\%", height = "400px",
  click = "", dblclick = "", hover = "", hoverDelay = "",
  hoverDelayType = "", brush = "", clickId = "", hoverId = "",
  inline = FALSE)
}
\arguments{
\item{Id}{outputId内码}

\item{width}{宽度}

\item{height}{高度}

\item{click}{单击}

\item{dblclick}{双击}

\item{hover}{悬停}

\item{hoverDelay}{延时悬停}

\item{hoverDelayType}{延时悬停类型}

\item{brush}{刷子}

\item{clickId}{单击内码}

\item{hoverId}{悬停内码}

\item{inline}{是否一行}
}
\value{
返回值
}
\description{
定义output.image实例化入口函数
}
\examples{
output.image();
}
