% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.print.R
\docType{class}
\name{output.print-class}
\alias{output.print-class}
\title{定义output.print基类,可以实例化}
\value{
返回值
}
\description{
定义output.print基类,可以实例化
}
\section{Slots}{

\describe{
\item{\code{placeholder}}{logical. 点位符}
}}

\examples{
output.print();
}
