% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.print.R
\name{output.print}
\alias{output.print}
\title{定义output.print的同名入口函数}
\usage{
output.print(Id = "verbatimTextOutput1", placeholder = FALSE)
}
\arguments{
\item{Id}{outputId}

\item{placeholder}{占位符}
}
\value{
返回值
}
\description{
定义output.print的同名入口函数
}
\examples{
output.print()
}
