% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.text.R
\name{output.text}
\alias{output.text}
\title{定义output.text的同名入口函数}
\usage{
output.text(Id = "textOutput1", inline = FALSE)
}
\arguments{
\item{Id}{outputId}

\item{inline}{是否同一行}
}
\value{
返回值
}
\description{
定义output.text的同名入口函数
}
\examples{
output.text();
}
