% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.button.download.R
\name{process.button.download}
\alias{process.button.download}
\title{定义相应的处理}
\usage{
process.button.download(input, output, session, data,
  filename = "下载文件.xlsx")
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{全文}

\item{data}{数据}
}
\value{
返回值
}
\description{
定义相应的处理
}
\examples{
process.button.download();
}
