% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.dataTable.preview.R
\name{process.dataTable.preview}
\alias{process.dataTable.preview}
\title{dataTable预览通用逻辑处理}
\usage{
process.dataTable.preview(input, output, session, lengthMenu = c(10, 50,
  100, 1000, 10000), pageLength = 10)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{lengthMenu}{每页显示条目数}

\item{pageLength}{默认条目数}

\item{data}{数据}
}
\value{
返回值
}
\description{
dataTable预览通用逻辑处理
}
\examples{
process.dataTable.preview();
}
