% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.image.R
\name{process.image}
\alias{process.image}
\title{定义相应的的处理}
\usage{
process.image(input, output, session, imageFile)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{imageFile}{imageFileName}
}
\value{
return value of a list
}
\description{
定义相应的的处理
}
\examples{
process.image();
}
