% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.plot.R
\name{process.plot}
\alias{process.plot}
\title{设置process.plot的处理函数}
\usage{
process.plot(input, output, session, plot_obj = quote(hist(faithful[, 2],
  breaks = 30, col = "darkgray", border = "white")))
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{plot_obj}{plot obj expression}
}
\value{
return value
}
\description{
设置process.plot的处理函数
}
\examples{
process.plot();
}
