% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.print.R
\name{process.print}
\alias{process.print}
\title{定义相应的处理数据}
\usage{
process.print(input, output, session, data = letters)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会议}

\item{data}{数据}
}
\value{
返回值
}
\description{
定义相应的处理数据
}
\examples{
process.print();
}
