% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.dataTable.R
\name{run.dataTable.preview2}
\alias{run.dataTable.preview2}
\title{将run.dataTable定义的相应的后台组件进行处理}
\usage{
run.dataTable.preview2(Id)
}
\arguments{
\item{Id}{域名空间Id}
}
\value{
返回值
}
\description{
将run.dataTable定义的相应的后台组件进行处理
}
\examples{
run.dataTable.preview2();
}
