% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb_shiny_input.R
\docType{methods}
\name{shiny_input,ui_date-method}
\alias{shiny_input,ui_date-method}
\title{shiny_input for ui_date}
\usage{
\S4method{shiny_input}{ui_date}(x)
}
\arguments{
\item{x}{ui_date.  对象}
}
\value{
返回值
}
\description{
shiny_input for ui_date
}
\examples{
shiny_input(x);
}
