% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb_shiny_input.R
\docType{methods}
\name{shiny_input,ui_number_slider-method}
\alias{shiny_input,ui_number_slider-method}
\title{定义shiny_input在ui_number_slider上的实现代码}
\usage{
\S4method{shiny_input}{ui_number_slider}(x)
}
\arguments{
\item{x}{ui_number_slider. ui_numer_slider实例}
}
\value{
返回值
}
\description{
定义shiny_input在ui_number_slider上的实现代码
}
\examples{
shiny_input(bb);
}
