% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb_shiny_input.R
\docType{methods}
\name{shiny_input,ui_password-method}
\alias{shiny_input,ui_password-method}
\title{定义shiny_input对ui_password的实现}
\usage{
\S4method{shiny_input}{ui_password}(x)
}
\arguments{
\item{x}{ui_password. 实例化对象ui_password}
}
\value{
返回值
}
\description{
定义shiny_input对ui_password的实现
}
\examples{
shiny_input(x);
}
