% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb_shiny_input.R
\docType{methods}
\name{shiny_input,ui_text-method}
\alias{shiny_input,ui_text-method}
\title{shiny_input用于ui_text的处理方法定义}
\usage{
\S4method{shiny_input}{ui_text}(x)
}
\arguments{
\item{x}{ui_text.}
}
\value{
返回值
}
\description{
shiny_input用于ui_text的处理方法定义
}
\examples{
shiny_input(bb);
}
