% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyElement.R
\docType{class}
\name{shinyElement-class}
\alias{shinyElement-class}
\title{定义shinyElement的虚拟类}
\value{
返回值
}
\description{
定义shinyElement的虚拟类
}
\section{Slots}{

\describe{
\item{\code{Id}}{character. 内码Id}
}}

\examples{
shinyElement;
}
