% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInputElement.R
\docType{class}
\name{shinyInputElement-class}
\alias{shinyInputElement-class}
\title{定义类shinyInputUI,不用于实例化}
\value{
返回值
}
\description{
定义类shinyInputUI,不用于实例化
}
\section{Slots}{

\describe{
\item{\code{label}}{character.  标题名}

\item{\code{width}}{character.  宽度}
}}

\examples{
shinyInputElement();
}
