% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyOutputElement.R
\docType{class}
\name{shinyOutputElement-class}
\alias{shinyOutputElement-class}
\title{定义类shinyOutputUI,不用于实例化}
\value{
返回值
}
\description{
定义类shinyOutputUI,不用于实例化
}
\examples{
shinyOutputElement();
}
