% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superButton.R
\docType{class}
\name{superButton-class}
\alias{superButton-class}
\title{定义superButton虚拟类}
\value{
返回值
}
\description{
定义superButton虚拟类
}
\section{Slots}{

\describe{
\item{\code{icon}}{character. 图标}

\item{\code{...}}{list.  更新参数}
}}

\examples{
superButton
}
