% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superDate.R
\docType{class}
\name{superDate-class}
\alias{superDate-class}
\title{定义superDate超级类,不要实例化}
\value{
返回值
}
\description{
定义superDate超级类,不要实例化
}
\section{Slots}{

\describe{
\item{\code{min}}{Date. 最小日期}

\item{\code{max}}{Date.  最大日期}

\item{\code{format}}{character. 日期格式}

\item{\code{startview}}{character. 展示形式}

\item{\code{weekstart}}{integer. 一周开始于}

\item{\code{language}}{character.  英文}

\item{\code{autoclose}}{logical. 自动关闭}
}}

\examples{
superDate();
}
