% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superImagePlot.R
\docType{class}
\name{superImagePlot-class}
\alias{superImagePlot-class}
\title{定义superImagePlot超级类，不支持实例化}
\value{
返回值
}
\description{
定义superImagePlot超级类，不支持实例化
}
\section{Slots}{

\describe{
\item{\code{width}}{character. 宽度}

\item{\code{height}}{character. 高度}

\item{\code{click}}{character. 单击}

\item{\code{dblclick}}{character. 双击}

\item{\code{hover}}{character.  悬停}

\item{\code{hoverDelay}}{character.  延时悬停}

\item{\code{hoverDelayType}}{character. 延时类型}

\item{\code{brush}}{character.    刷子}

\item{\code{clickId}}{character.  单击ID}

\item{\code{hoverId}}{character.  双击ID}

\item{\code{inline}}{logical.    是否一行还是多选}
}}

\examples{
superImagePlot();
}
