% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superMultiOption.R
\docType{class}
\name{superMultiOption-class}
\alias{superMultiOption-class}
\title{定义superMultiOption超级类,不可实例化}
\value{
返回值
}
\description{
定义superMultiOption超级类,不可实例化
}
\section{Slots}{

\describe{
\item{\code{choiceNames}}{list. 列表可以是文字或icon}

\item{\code{choiceValues}}{list. 必须是文字列表}

\item{\code{selected}}{character.  默认选项}

\item{\code{inline}}{logical.     按行展示}
}}

\examples{
 superMultiOption();
}
