% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superNumeric.R
\docType{class}
\name{superNumeric-class}
\alias{superNumeric-class}
\title{定义数值型的超级类，不可实例化}
\value{
返回值
}
\description{
定义数值型的超级类，不可实例化
}
\section{Slots}{

\describe{
\item{\code{value}}{numeric. 默认值}

\item{\code{min}}{numeric.   最大值}

\item{\code{max}}{numeric.   最小值}

\item{\code{step}}{numeric.  步长}
}}

\examples{
superNumeric();
}
