% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superText.R
\docType{class}
\name{superText-class}
\alias{superText-class}
\title{定义superText超级类}
\value{
返回值
}
\description{
定义superText超级类
}
\section{Slots}{

\describe{
\item{\code{value}}{character.  默认值}

\item{\code{placeholder}}{character. 点位符}

\item{\code{isPassword}}{是否为密码}
}}

\examples{
superText();
}
