% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superTextPlusUI.R
\docType{class}
\name{superTextPlusUI-class}
\alias{superTextPlusUI-class}
\title{定义superTextPlusUI超级类,不可实例化}
\value{
返回值
}
\description{
定义superTextPlusUI超级类,不可实例化
}
\section{Slots}{

\describe{
\item{\code{inline}}{logical. 是否同一行，横向}
}}

\examples{
superTextPlusUI
}
