% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_defineClasses.R
\docType{class}
\name{ui_date-class}
\alias{ui_date-class}
\title{定义日期型的输入控件ui_date S4类}
\value{
返回值
}
\description{
定义日期型的输入控件ui_date S4类
}
\section{Slots}{

\describe{
\item{\code{inputId}}{character.内码}

\item{\code{label}}{character. 标签}

\item{\code{value}}{character.  默认值}

\item{\code{min}}{character. 最小值}

\item{\code{max}}{character. 最大值}

\item{\code{format}}{character. 日期格式}

\item{\code{startview}}{character. 日历形式month,year,decade}

\item{\code{weekstart}}{integer. 0表示周日开始，1表示周1开始}

\item{\code{language}}{character. 语言en,zh-CN,zh-TW}

\item{\code{width}}{character. 宽度}

\item{\code{autoclose}}{logical. 是否自动关闭}
}}

\examples{
ui_date(x);
}
