% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_defineClasses.R
\name{ui_date}
\alias{ui_date}
\title{ui_date实例化入口func}
\usage{
ui_date(inputId = "date1", label = "输入日期:", value = "2016-01-01",
  min = "1984-01-01", max = "2100-12-31", format = "yyyy-mm-dd",
  startview = "month", weekstart = 0L, language = "zh-CN",
  width = "100\%", autoclose = TRUE)
}
\arguments{
\item{inputId}{内码}

\item{label}{标签}

\item{value}{默认值}

\item{min}{最小值}

\item{max}{最大值}

\item{format}{日期格式}

\item{startview}{日历格式}

\item{weekstart}{开始于}

\item{language}{日历语言}

\item{width}{宽度}
}
\value{
返回值
}
\description{
ui_date实例化入口func
}
\examples{
ui_date();
}
