% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_defineClasses.R
\docType{class}
\name{ui_file-class}
\alias{ui_file-class}
\title{Title}
\value{
返回值
}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{inputId}}{character. 内码}

\item{\code{label}}{character.  标签}

\item{\code{multiple}}{logical. 是否支持批量上传}

\item{\code{accept}}{character.  是否指定文件后缀名}

\item{\code{width}}{character.   宽度}

\item{\code{buttonLabel}}{character.   按纽文字内容}

\item{\code{placeholder}}{character.   占位符内容}
}}

\examples{
 ui_file("sample.xlsx")
}
