% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_defineClasses.R
\name{ui_file}
\alias{ui_file}
\title{ui_file上传控件的入口实例化函数}
\usage{
ui_file(inputId = "fileInput1", label = "上传文件:",
  multiple = FALSE, accept = "", width = "100\%",
  buttonLabel = "浏览...",
  placeholder = "没有选中任何文件！")
}
\arguments{
\item{inputId}{内码}

\item{label}{标签}

\item{multiple}{是否批量上传}

\item{accept}{是否指定文件后缀名}

\item{width}{是否指定宽度}

\item{buttonLabel}{按纽标签名称}

\item{placeholder}{点位符内容}
}
\value{
返回值
}
\description{
ui_file上传控件的入口实例化函数
}
\examples{
 ui_file();
}
