% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_defineClasses.R
\docType{class}
\name{ui_number_slider-class}
\alias{ui_number_slider-class}
\title{定义新的ui_number_slider option选英}
\value{
返回值
}
\description{
定义新的ui_number_slider option选英
}
\section{Slots}{

\describe{
\item{\code{inputId}}{character.  内码}

\item{\code{label}}{character.  标签名称}

\item{\code{max}}{numeric.  最大值}

\item{\code{min}}{numeric.  最小值}

\item{\code{value}}{numeric. 默认值}

\item{\code{step}}{numeric. 步长}

\item{\code{width}}{character. 宽度}
}}

\examples{
ui_number_slider
}
