% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_defineClasses.R
\name{ui_number_slider}
\alias{ui_number_slider}
\title{定义使用创建ui_number_input实例化函数}
\usage{
ui_number_slider(inputId = "sliderInput1", label = "number slider1",
  min = 1, max = 500, value = 300, step = 5, width = "400px")
}
\arguments{
\item{inputId}{内码}

\item{label}{标签}

\item{min}{最小值}

\item{max}{最大值}

\item{value}{默认值}

\item{step}{步长}

\item{width}{宽度}
}
\value{
返回值
}
\description{
定义使用创建ui_number_input实例化函数
}
\examples{
ui_number_slider();
}
