% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_defineClasses.R
\docType{class}
\name{ui_password-class}
\alias{ui_password-class}
\title{创建日期控件的选项}
\value{
还回值
}
\description{
创建日期控件的选项
}
\section{Slots}{

\describe{
\item{\code{inputId}}{character. 内码}

\item{\code{label}}{character.  标签}

\item{\code{value}}{character. 初始值}

\item{\code{width}}{character. 宽度}
}}

\examples{
ui_password(aa);
}
