% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_defineClasses.R
\name{ui_password}
\alias{ui_password}
\title{ui_password的实例化function}
\usage{
ui_password(inputId = "password1", label = "输入密码:",
  value = "888888", width = "100\%")
}
\arguments{
\item{inputId}{内码}

\item{label}{标签}

\item{value}{初始值}

\item{width}{控件宽度}
}
\value{
返回值
}
\description{
ui_password的实例化function
}
\examples{
ui_password();
}
