% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_defineClasses.R
\docType{class}
\name{ui_text-class}
\alias{ui_text-class}
\title{定义ui_text iption的特殊类，用于存储相应的界面数据}
\value{
返回值
}
\description{
定义ui_text iption的特殊类，用于存储相应的界面数据
}
\section{Slots}{

\describe{
\item{\code{inputId}}{character. 内码Id,用于input$inputId}

\item{\code{label}}{character.   标签}

\item{\code{value}}{character.   默认值}

\item{\code{width}}{character.  文本框宽度，默认为100%或400px}
}}

\examples{
ui_text("aaaa")
}
