% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_defineClasses.R
\name{ui_text}
\alias{ui_text}
\title{定义一个实例化函数，用于数据存储}
\usage{
ui_text(inputId = "txt1", label = "text1_caption", value = "",
  width = "100\%")
}
\arguments{
\item{inputId}{标签内码}

\item{label}{标签标题}

\item{value}{默认值}

\item{width}{宽度}
}
\value{
返回值
}
\description{
定义一个实例化函数，用于数据存储
}
\examples{
bb <-ui_text(inputId = 'aa',label = "bb",value = 'cc',width = "80\%");
}
